Refactor body node utils
