+++ src/Nodes/Body.ts
@@ -6,7 +6,7 @@
 
 import { Node, NodeViewRendererProps, mergeAttributes } from "@tiptap/core";
 import { BODY_ATTRIBUTES, BODY_NODE_NAME } from "../constants/body";
-import { isBodyNode } from "../utils/pageRegion/body";
+import { isBodyNode } from "../utils/nodes/body";
 import { addNodeAttributes, parseHTMLNode } from "../utils/nodes/node";
 import { calculateShorthandMargins, calculateBodyMargins } from "../utils/pageRegion/margins";
 import { mm } from "../utils/units";
